Revert "comitPersonalCambiosPedidos"
This reverts commit 774d933742a4479a05577e8938282cda943b2fc0.
