+++ RicoProject/src/Archivos/ManejadordeArchivos.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
@@ -54,11 +53,7 @@ public class ManejadordeArchivos {
 	{
 		ListadoPedidos destinoPedido = new ListadoPedidos();
 		destinoPedido = pedido.leerPedidos();
-		if(!destinoPedido.getListaPedidosContenedorMap().containsKey(definirFecha()))
-		{
-			destinoPedido.inciarDia();
-		}
-		
+		destinoPedido.inciarDia();
 		return destinoPedido;	
 	}
 	
@@ -188,19 +183,6 @@ public class ManejadordeArchivos {
 			productos.agregarCombo(lista);
 		}
 		
-	}
-	
-	public static String definirFecha()
-	{
-		Date fechaOrigen = new Date();
-		int dia,mes,anio;
-		String strFecha;
-		dia=fechaOrigen.getDate();
-		mes=fechaOrigen.getMonth()+1;
-		anio=fechaOrigen.getYear()+1900;
-		strFecha = dia+"/"+mes+"/"+anio;
-		
-		return strFecha;
 	}
 	
 }
